+++ src/players.c
@@ -911,7 +911,9 @@ int load_char(const char *name, struct char_data *ch)
         {
           if (GET_ARCANE_MARK(ch))
             free(GET_ARCANE_MARK(ch));
-          GET_ARCANE_MARK(ch) = strdup(line);
+          GET_ARCANE_MARK(ch) = NULL;
+          if (*line && strcmp(line, "(null)") && strcmp(line, "null"))
+            GET_ARCANE_MARK(ch) = strdup(line);
         }
         break;
 
@@ -2550,8 +2552,8 @@ bool save_char_checked(struct char_data *ch, int mode)
 
   if (GET_HOST(ch))
     BUFFER_WRITE("Host: %s\n", GET_HOST(ch));
-  // save arcane mark no matter what
-  BUFFER_WRITE("AMrk: %s\n", GET_ARCANE_MARK(ch));
+  if (GET_ARCANE_MARK(ch) && *GET_ARCANE_MARK(ch))
+    BUFFER_WRITE("AMrk: %s\n", GET_ARCANE_MARK(ch));
   if (GET_HEIGHT(ch) != PFDEF_HEIGHT)
     BUFFER_WRITE("Hite: %d\n", GET_HEIGHT(ch));
   if (HIGH_ELF_CANTRIP(ch))